fussin

+++ frb/builds/build_hosts.py
@@ -301,10 +301,11 @@ def run(host_input:pandas.core.series.Series,
                 merge_tbl['Name'] = file_root
 
     # Remove NSC for now
-    for key in merge_tbl.keys():
-        if 'NSC' in key:
-            merge_tbl.remove_column(key)
-            print(f"Removing NSC column: {key}")
+    if merge_tbl is not None:
+        for key in merge_tbl.keys():
+            if 'NSC' in key:
+                merge_tbl.remove_column(key)
+                print(f"Removing NSC column: {key}")
     # Finish
     if merge_tbl is not None:
         # Dust correct
